replace globals with real types
This commit introduces u2ftoken.Token, which accepts a *storage.Storage
as its sole field.
We now provide a sane way for callers to initialize a U2F token with a
given storage, instead of relying on storage's global variables.

Also now u2fhid.Handler is not a shared variable anymore, and it's
caller responsibility to properly initialize it.